+++ web/tailwind.config.ts
@@ -4,7 +4,14 @@ const config: Config = {
   darkMode: "class",
   content: ["./app/**/*.{ts,tsx}"],
   theme: {
-    extend: {},
+    extend: {
+      keyframes: {
+        indeterminate: {
+          "0%": { transform: "translateX(-100%)" },
+          "100%": { transform: "translateX(350%)" },
+        },
+      },
+    },
   },
   plugins: [],
 };